refactor accessory out of Avatar
